cleared the directory and added the comments in the main_interface file
